+++ decidim-participatory_processes/app/helpers/decidim/participatory_processes/participatory_process_helper.rb
@@ -21,6 +21,22 @@ module Decidim
         dates = [participatory_process_step.start_date, participatory_process_step.end_date]
         dates.map { |date| date ? localize(date.to_date, format: :default) : "?" }.join(" - ")
       end
+
+      # Public: Builds the URL for the step Call To Action. Takes URL params
+      # into account.
+      #
+      # process - a ParticipatoryProcess
+      #
+      # Returns a String that can be used as a URL.
+      def step_cta_url(process)
+        base_url, params = decidim_participatory_processes.participatory_process_path(process).split("?")
+
+        if params.present?
+          [base_url, "/", process.active_step.cta_path, "?", params].join("")
+        else
+          [base_url, "/", process.active_step.cta_path].join("")
+        end
+      end
     end
   end
 end